[coordinates] added in memory mixin, moved workers from Iterable to DataSource
